Update Polars to v0.18.0
This is a minor update but since Polars' API is unstable, it also breaks
some of Explorer's APIs. Most notable are the functions from
`Explorer.Series` module (and the Series behaviour).

The changes are related to `Explorer.Series.rolling_` functions.
They have changed to accept `min_periods` (int) and `center` (bool).
The `ignore_nil?` option is gone.

For future reference, this is the diff of commits between Polars'
versions in this update:

https://github.com/pola-rs/polars/compare/2055cf205e992c5eb0809ada55e067ef4c293b0d...rust-polars-0.18.0