Use `rustler_precompiled` for releases
It makes Explorer force the usage of precompiled NIFs when it's not in a
pre-release version. Therefore people don't need to have Rust installed.

Users still can force the build by setting the system env var:

    EXPLORER_BUILD=true mix compile --force

Or by setting an application env:

    config :rustler_precompiled, :force_build, explorer: true

This is related to https://github.com/elixir-nx/explorer/pull/122