Always build NIF on CI
This change makes sure that we always build the NIF on CI. Even if we
are releasing an official version, we should build from source.
This will prevent the build to fail in those occasions, where the
precompiled NIFs won't be available.

This change also include a restriction for running the "release"
workflow - that builds the precompiled NIFs - only if the "native" code
changed. This is because we almost don't depend on Elixir code for that
workflow.