Keep track of names & dtypes and introduce a "virtual" out DF (#238)
The idea of this change is to simplify the implementation of data frame backends
by applying the changes to a "virtual DF" and letting the backend decide how
to apply those changes.

This is the list of functions affected:

- select
- mutate
- rename
- pivot_longer
- group_by
- ungroup
- summarise
- join
- distinct

Detailed changes:

* WIP: add `Explorer.Backend.DataFrame.new/3`

This function is meant to create a new df with a backend

* WIP: simplify the `select` callback apply changed df

The idea is to apply an operation in memory, and then apply in the
backend. We want to make easier for the backend to apply the select
operation.

* Remove shared function for columns update

* Fix mistakes after rebase

* Simplify the option "on" for DF backend callback

This makes easier to implement this function for a new backend because
people don´t need to treat the column as string.

* Remove branch for treating lazy frame when creating df

* Add more tests to `DF.select/3`

* Change rename callback to accept a separated arg as out df

This is needed in order to make easier for devs to implement backends.

* Change `pivot_longer` callback to receive "out df"

This makes the callback simpler by removing the need for some arguments

* Refactor `pivot_longer` to use MeltArgs and avoid rename

This is a simplification of the operation at the Elixir side, because we
now have `melt2`:

https://pola-rs.github.io/polars/polars/frame/struct.DataFrame.html#method.melt2

* Improve `pivot_wider` tests and callback

The callback should receive only one `names_from` and `values_from`.

* Add one more test for pivot_wider

* Add a test for custom "on" - DF.join

* Refactor "group_by" to use helpers

* Normalize `group_by` to accept columns as atoms

Also validate it.

* WIP: add a test to "mutate" with groups

Also this fixes a small but that was preventing data from groups to be
pulled from the "groups DF".

And also I added some question points.

* WIP: making the `mutate` receive an "out_df"

This out_df will have the new dtypes and will replace the DF at the end
of the operation

* Refactor `PolarsBackend.DataFrame.mutate` to accept "out_df"

This change makes usage of the `out_df`, with updated names and dtypes.

* Rename helper function and use in places we need

* WIP: making "group_by" receive "out_df"

* Simplify some parts of polars backend and DF

* Make "ungroup" receive "out_df"

* Check that names where updated in the DFs

* Add one more example to "group_by"

* Apply on groups function

This will make clear of what is happening on grouped DFs.

Inspired by https://docs.rs/polars/0.21.1/polars/frame/groupby/struct.GroupBy.html#method.apply

* WIP: Add more tests

* Add tests to summarise

* Using "out_df" for summarise

* Fix the callback of summarise

* Fix bug with "join" operation when type is "right"

* Add "out_df" to "DF.join" callback

This makes the join operation receive the projection of the output DF of
joining the two data frames.

* Make "distinct" use "out_df"

* Using `df.names` and `df.dtypes` instead of functions

* Fix some TODOs related to Polars Backend

* Simplify mutate and keep columns order

* Minor improvements

* Replace MapSet by Enum.uniq

* Minor typo

* Sort DF after apply on groups

This makes the DF keep the original ordering after apply a callback to a
group.