Refactor PolarsBackend.Series to improve ops using scalar values (#370)
* Refactor Series backend to improve ops using scalar values

This is a refactor in the PolarsBackend.Series to reduce memory usage by
avoiding the need to create series for scalar values. Instead it is using
the lazy backend and expressions (lazy series) to pass down the
literal value to Rust.

* Ensure that we use Polars DF

This is going to prevent problems if the user configure different
backends for series and dataframes.