WIP: initial work to add support for list dtypes
This is work-in-progress and adds the bases of "list" dtypes.

It also adds the `to_list/1` lazy operation that takes a series
and creates a list series with the elements of that series.
This is useful for aggregations, when you want to capture the elements
of a given group. Eg.:

```elixir
df = Explorer.DataFrame.new(a: [1, 1, 2, 2], b: [9, 8, 7, 6])

grouped = Explorer.DataFrame.group_by(df, :a)

Explorer.DataFrame.summarise_with(grouped, fn df -> [b_merged: Explorer.Series.to_list(df["b"])] end)
```

The result is going to be something like this:

      #Explorer.DataFrame<
        Polars[2 x 2]
        a integer [1, 2]
        b_merged list(integer) [[ 9, 8 ], [ 7, 6 ]]
      >

Related to:
- https://github.com/elixir-nx/explorer/issues/296
- https://github.com/elixir-nx/explorer/issues/400