Fix pivot wider if the series of column names contains nil
This is going to correct name the column as "nil" instead of breaking.

There is a potential caveat that is when a dataframe contains already a
"null" column, and is going to create another "null" column. This is
going go break.
Since that case is less plausable, we are ignoring it.