Fix some `DF.pivot_wider/3` bugs (#535)
* Fix pivot wider if the series of column names contains nil

This is going to correct name the column as "nil" instead of breaking.

There is a potential caveat that is when a dataframe contains already a
"null" column, and is going to create another "null" column. This is
going go break.
Since that case is less plausable, we are ignoring it.

* Remove "out_df" from pivot_wider and impl in Rust

Since we don't have a "lazy" counterpart, it's easier to
implement the function without the "out_df".

* Handle possible duplicated columns after pivot_wider

This is an attempt to solve the problem of having multiple columns with
the same name after pivoting wider. This should probably happen if a new
column is equal to an ID column.

* Add one more case

* Filter out float columns from id_columns selection

* Fix filter of groups that are kept after pivot wider

Also change the place we manipulate that to be at the DF level,
not the backend.

Add docs as well.