Improve caching on CI workflows (#605)
The cached Elixir build is now taking around 2min30s, against
the non-cached that is taking around 8 min. The simplification
in the cached paths makes harder to have a miss.
So we should have quicker suite.

The precompilation workflow now takes around 18 minutes,
compared to the 30 minutes from the version without any cache.