Change compression backend for Polars
This is needed because of an odd problem we got recently after the
update of Rust Nightly version: the existing backend - `miniz_oxide`,
is breaking the BEAM because it is using a stack size that is greater
than the default value.

We have to manually change the size at boot time to make it work.

Using `ELIXIR_ERL_OPTIONS="+sssdcpu 128"` does the trick, but is not
very flexible.