Add necessary cargo features
These features are useful to control which NIF version we are going to compile
the project to, and which paths are going to be disabled due to incompatibility
with some targets we support.

This is using most of the work that is present in https://github.com/philss/rustler-precompiled-action/pull/4
But with an additional "cargo-args" that is experimental.