Fix FSS.S3 support when bucket is not provided
This makes Explorer treat the "endpoint" as already containing the bucket
name, and force the virtual host style in order to make Polars ignore the
bucket name.

Since FSS is always building the correct endpoint in case the user does not
provide, we changed the endpoint to be required in the Rust struct.

This work is related to https://github.com/elixir-explorer/fss/pull/2