Change `:float` to `{:f, 64}` and add `{:f, 32}` dtype
The idea is to support more dtypes in the way they are represented
in Arrow/Polars.

In the near future, we won't need to "normalise" numeric columns anymore,
and by consequence, Explorer is going to use less memory if needed.

This is related to https://github.com/elixir-explorer/explorer/issues/471