Add `re_contains/2` and `re_replace/3` to match with a regex
The regular expression must follow the `regex` crate rules, and should
be a string.

The idea of keeping them separate is because we cannot use the Elixir
standard regex, and we need to use strings for both literal and regex
versions. To make the difference explicit, we are using the `re_` prefix
for functions that accept regexes as strings.

This is related to https://github.com/elixir-explorer/explorer/issues/353