Improve performance of "concat_columns" (#938)
* Improve implementation of concat_columns

The idea is to optimize https://github.com/elixir-explorer/explorer/issues/937

* Improve performance of "concat_columns" for lazy frames

This is a different approach that consumes less memory and is faster.
There is a small change in behaviour, which is the following: now we can
concat DFs of different heights. The smaller columns will be filled with
`nil` values.