Implement Property Tests for `DataFrame.new` (#1012)
* Implement Property Tests for DataFrame.new

Co-authored-by: Billy Lanchantin <william.lanchantin@cargosense.com>

* First pass at changes

* Replace old generators

The existing property test used a special case
of the new generator logic. This replaces that
special case with the new generator.

* Add (skipped) serialization property tests

We should be able to serialize any DataFrame
or document the cases where we can't. None of
these are working at the moment.

* Drop max_runs to 1_000

* Fix bad copy/paste

* Add notes

* Switch to the `test_type:property` tag

Before we weren't actually skipping property
tests. We were excluding the`property`tag, but
that tag wasn't actually being set.

* Whoops! Revert that last change

Turns out we _were_ setting the `property` tag.

---------

Co-authored-by: Billy Lanchantin <william.lanchantin@cargosense.com>