Use `coef: i128` in decimal (#1017)
* Bump up the number of bits from 64 to 128

Before we were using `coef: u64` as that was the
largest type that was also guarunteed to be
positive. This left half the bits unused, however.
This change makes it so we're using all available
bits. Even though technically we could make an
invalid ExDecimal after this change, it's not a
practical concern.

* Increase precision max to 38 in generator

This reverts an old change that was a work around
for the `coef: u64` thing.

* Remove misleading comment

* Fix linting