Change how (naive) datetimes are encoded
Before we were always converting first to a
microsecond-based representation then to the
final representation. The intermediate conversion
is unecessary and risks overflows when trying to
convert to a different time unit later. This approach
converts directly to i64 from the Elixir struct and
time unit.

The drawback is that we can no longer build i64
datetimes directly from Elixir structs.