Fix datetime encoding overflow (#1020)
* Change how (naive) datetimes are encoded

Before we were always converting first to a
microsecond-based representation then to the
final representation. The intermediate conversion
is unecessary and risks overflows when trying to
convert to a different time unit later. This approach
converts directly to i64 from the Elixir struct and
time unit.

The drawback is that we can no longer build i64
datetimes directly from Elixir structs.

* Allow fine-grained timestamp generation

* Restrict nanosecond datetimes to a set
  which is representable by i64
* Make datetime generation funs public

* Add properties

* Test new timestamp conversion error type

* Use correct module

I'm surprised this worked...

* Fix linting

* `format!("...")` to `"...".to_string()`
* `.map_or(...)` to `.ok_or`
* `x.clone()` to `*x`
* `.map(|x| Some(x))` to `.map(Some)`

* Turns out the `_res` thing is weird

This change reverts that. It also refactors the
change in `series.rs` to be more readable.