chore: update flake.nix
I believe the minimum rust version was incremented a few months ago, but
the flake inputs were not updated as well. I was getting an error about
a failure with cargo metadata and the lockfile version being 4.

Now it all works. It also updates the Erlang version 27, I assume that
is fine.